Extract CommitMessage into a separate file
I'm doing this because I want to start adding to the implementation.
CommitMessage should probably live in a different crate as it's shared
by multiple applications. This seems like a reasonable first step though.
