+++ pb-commit-message-lints/src/lints/lib/mod.rs
@@ -1,0 +1,3 @@
+pub mod commit_message;
+
+pub use commit_message::CommitMessage;
