+++ pb-commit-message-lints/src/lints/mod.rs
@@ -1,7 +1,5 @@
 use std::fmt::Display;
 
-use regex::Regex;
-
 use crate::{
     errors::PbCommitMessageLintsError,
     external::vcs::Vcs,
@@ -12,102 +10,10 @@ use crate::{
         Lints::{DuplicatedTrailers, JiraIssueKeyMissing, PivotalTrackerIdMissing},
     },
 };
-use std::{convert::TryFrom, fs::File, io::Read, path::PathBuf};
 
-pub struct CommitMessage {
-    contents: String,
-}
+pub mod lib;
 
-impl CommitMessage {
-    #[must_use]
-    pub fn new(contents: String) -> CommitMessage {
-        CommitMessage { contents }
-    }
-
-    pub fn matches_pattern(&self, re: &Regex) -> bool {
-        re.is_match(&self.contents)
-    }
-
-    #[must_use]
-    pub fn get_trailer(&self, trailer: &str) -> Vec<&str> {
-        self.contents
-            .lines()
-            .filter(|line: &&str| CommitMessage::line_has_trailer(trailer, line))
-            .collect::<Vec<_>>()
-    }
-
-    fn line_has_trailer(trailer: &str, line: &str) -> bool {
-        line.starts_with(&format!("{}:", trailer))
-    }
-}
-
-impl TryFrom<PathBuf> for CommitMessage {
-    type Error = PbCommitMessageLintsError;
-
-    fn try_from(value: PathBuf) -> Result<Self, Self::Error> {
-        let mut file = File::open(value)?;
-        let mut buffer = String::new();
-
-        file.read_to_string(&mut buffer)
-            .map_err(PbCommitMessageLintsError::from)
-            .map(move |_| CommitMessage::new(buffer))
-    }
-}
-
-impl Display for CommitMessage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.contents)
-    }
-}
-
-#[cfg(test)]
-mod test_commit_message {
-    use pretty_assertions::assert_eq;
-    use regex::Regex;
-
-    use crate::lints::CommitMessage;
-
-    #[test]
-    fn with_trailers() {
-        let commit = CommitMessage::new(
-            r#"Some Commit Message
-
-Anything: Some Trailer
-Anything: Some Trailer
-Another: Trailer
-"#
-            .into(),
-        );
-
-        assert_eq!(vec!["Another: Trailer"], commit.get_trailer("Another"));
-        assert_eq!(
-            vec!["Anything: Some Trailer", "Anything: Some Trailer"],
-            commit.get_trailer("Anything")
-        )
-    }
-
-    #[test]
-    fn regex_matching() {
-        let commit = CommitMessage::new(
-            r#"Some Commit Message
-
-Anything: Some Trailer
-Anything: Some Trailer
-Another: Trailer
-"#
-            .into(),
-        );
-
-        assert_eq!(
-            true,
-            commit.matches_pattern(&Regex::new("[AB]nything:").unwrap())
-        );
-        assert_eq!(
-            false,
-            commit.matches_pattern(&Regex::new("N[oO]thing:").unwrap())
-        );
-    }
-}
+use lib::CommitMessage;
 
 /// The lints that are supported
 #[derive(Debug, Eq, PartialEq, Copy, Clone)]
